fix(daemon): unify live-sync client across CLI + tools; unregister stale registration on disabled create
- Extract DaemonSyncResult/trySyncToDaemon/SYNC_FAILED_NOTE into
  src/agent/daemon/http-client.ts; schedules.ts and schedule.ts both
  import from there, eliminating the duplicated silent-swallow copy.
- afk schedule add/remove/enable/disable now surface a warning when
  live-sync fails (was silently swallowed).
- createScheduleHandler disabled branch sends idempotent DELETE instead
  of a fake synced:true, so stale live registrations are unregistered.
- Update test assertion for disabled-create to match new syncDetail.
